Create convert.py
Checks model type of MODEL_PATH's model from .env and if is old ggml type, converts to ggjt and closes. If already ggjt, it just exits.

The code to check the model type is on line 863 "def lazy_load_file(path: Path) -> ModelPlus:"

Could be interesting if you could call it automatically from startllm and then after conversion change the name of your old file to oldmodel.bin and the new file to what you have in .env and then use it.